requirements.txt
scikit-learn
joblib
numpy
pandas
fastapi
uvicorn[standard]
pydantic

# Exporter for Google Cloud
opentelemetry-exporter-gcp-trace
